app/requirements.txt
setuptools
Flask == 2.3.*
flask-sqlalchemy == 3.1.*
mysqlclient == 2.0.*
flask-login
mod_wsgi
flask_wtf